fix: id필드를 int64로 변경

+++ pennyway-client-iOS/pennyway-client-iOS/Clean/Presentation/Scene/ChatRoomDetail/View/ChatRoomDetailView.swift
@@ -57,7 +57,7 @@ struct ChatRoomDetailView: View, ImageLoadable {
                     if let chatRoomId = chatRoom?.id {
                         Log.debug("[ChatRoomDetailView]: id까지 진입")
 
-                        viewModelWrapper.joinChatRoomViewModel.joinChatRoom(chatRoomId: Int64(chatRoomId), password: "") { success in
+                        viewModelWrapper.joinChatRoomViewModel.joinChatRoom(chatRoomId: chatRoomId, password: "") { success in
                             Log.debug("[ChatRoomDetailView]: joinChatRoom까지 진입")
 
                             if success {
